requirements.txt
uvicorn==0.35.0
fastapi==0.115.14
matplotlib==3.10.3
numpy==2.1.3
pandas==2.3.0
pillow==11.3.0
scikit-learn==1.7.1
tensorflow[and-cuda]==2.19.0
python-multipart==0.0.20

